fix load value of dynamic values of scripted module
The value of dynamic values (e.g. local.values.addIntParameter called by moduleParameterChanged depending on a parameter value) is saved but not loaded because the values do not exist yet, because the parameters are not loaded yet.
Loading parameters first will trigger the script's scriptParameterChanged that may be responsible for creating dynamic values. Then value of created values will load just fine.